highlight levels greater than 1 as in progress
I noticed a bug that occurs when I've completed a competency and there is a level 2 available for it. The level 1 will be green on my competencies page but the level 2 will be grey indicating I don't plan on learning it yet (however there will be no button to add it as something I want to learn because it's already on the sheet).

This change will add the `in-progress` class to a competency pill if the level is 0 OR if the current user's level is less than the competencies level but they have it on their sheet.